requirements.txt
fastapi~=0.115.12
uvicorn
torch~=2.6.0+cu118
torchvision~=0.21.0+cu118
pillow~=11.0.0
streamlit~=1.42.2
requests~=2.32.3

pydantic~=2.11.3
kagglehub~=0.3.12